Switched authToken requirement from Story to Feed
