+++ app/src/main/java/edu/byu/cs/tweeter/Client/presenter/FeedPresenter.java
@@ -6,6 +6,7 @@ import androidx.annotation.RequiresApi;
 
 import java.io.IOException;
 
+import byu.edu.cs.tweeter.shared.model.domain.AuthToken;
 import byu.edu.cs.tweeter.shared.net.TweeterRemoteException;
 import edu.byu.cs.tweeter.Client.model.services.FeedServiceProxy;
 import byu.edu.cs.tweeter.shared.request.FeedRequest;
@@ -22,5 +23,9 @@ public class FeedPresenter extends Presenter {
     @RequiresApi(api = Build.VERSION_CODES.O)
     public FeedResponse getFeed(FeedRequest request) throws IOException, TweeterRemoteException {
         return (new FeedServiceProxy()).getFeed(request);
+    }
+
+    public AuthToken findCurrentAuthToken() {
+        return getCurrentAuthToken();
     }
 }
