+++ shared/src/main/java/byu/edu/cs/tweeter/shared/request/FeedRequest.java
@@ -1,5 +1,6 @@
 package byu.edu.cs.tweeter.shared.request;
 
+import byu.edu.cs.tweeter.shared.model.domain.AuthToken;
 import byu.edu.cs.tweeter.shared.model.domain.Status;
 import byu.edu.cs.tweeter.shared.model.domain.User;
 
@@ -7,11 +8,17 @@ public class FeedRequest {
     private User user;
     private int limit;
     private Status lastStatus;
+    private AuthToken authToken;
 
-    public FeedRequest(User user, int limit, Status lastStatus) {
+    public FeedRequest(User user, int limit, Status lastStatus, AuthToken authToken) {
         this.user = user;
         this.limit = limit;
         this.lastStatus = lastStatus;
+        this.authToken = authToken;
+
+    }
+    public AuthToken getAuthToken() {
+        return authToken;
     }
 
     public User getUser() {
@@ -30,6 +37,8 @@ public class FeedRequest {
         user = null;
         limit = 0;
         lastStatus = null;
+        authToken = null;
+
     }
 
     public void setUser(User user) {
@@ -43,4 +52,9 @@ public class FeedRequest {
     public void setLastStatus(Status lastStatus) {
         this.lastStatus = lastStatus;
     }
+
+    public void setAuthToken(AuthToken authToken) {
+        this.authToken = authToken;
+    }
+
 }
